Moving the new `run_async` method to the `RayPipeline`
